Move all the commands to their own package internal/commands so they can be re-used by another binary.
