+++ cmd/docker-app/main.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"github.com/docker/app/internal"
+	app "github.com/docker/app/internal/commands"
 	"github.com/docker/cli/cli-plugins/manager"
 	"github.com/docker/cli/cli-plugins/plugin"
 	"github.com/docker/cli/cli/command"
@@ -10,7 +11,7 @@ import (
 
 func main() {
 	plugin.Run(func(dockerCli command.Cli) *cobra.Command {
-		return newRootCmd(dockerCli)
+		return app.NewRootCmd("app", dockerCli)
 	}, manager.Metadata{
 		SchemaVersion: "0.1.0",
 		Vendor:        "Docker Inc.",
