+++ internal/commands/init.go
@@ -1,4 +1,4 @@
-package main
+package commands
 
 import (
 	"github.com/docker/app/internal/packager"
