+++ internal/commands/inspect.go
@@ -1,4 +1,4 @@
-package main
+package commands
 
 import (
 	"github.com/deislabs/duffle/pkg/action"
