+++ internal/commands/parameters_test.go
@@ -1,4 +1,4 @@
-package main
+package commands
 
 import (
 	"testing"
